test: split comment action tests
